Refactor Artifact creation.
